synapse_backend/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
google-cloud-firestore
youtube-transcript-api
pydantic
python-multipart
python-dotenv
google-generativeai
google-cloud-storage
yt-dlp
